refact: move person + organization display to separate file
